fix(st): avoid enormous tf-a-stm32mp13*stm32
The current STM32MP13xx build produces tf-a-stm32mp13*stm32 which is
about 800 MiB is size. This is caused by default build-id emitted by
GNU linker and no-pie build. Change the linker flags such that the
build is static and no build-id is generated. This produces expected
binary, with the right size, which works on the actual hardware again.
This fix is similar to 72f4b70e8e8e ("fix(rcar-layout): fix tool build").

This failure is triggered using most of arm-linux-* toolchains, which
are generated by OE-core 5.0 Scarthgap, arm-linux-gnueabi* 15.2.0 in
Debian and such toolchain. The arm-none-eabi- toolchains do not seem
to trigger this.

Change-Id: I0f7d9f330dbe08d6adf491ed79b3a10a4adfecc7
Fixes: 6d26d75c374b ("fix(st): set no-pie option when building ST elf file")
Signed-off-by: Marek Vasut <marex@nabladev.com>